posix: potientially perform allocation in acquire_new_entry
This approach seems to be cleaner.

This routine is called from a single place only.
Having the allocation outside the routine raised such question
during code review:
"These asserts are only blown in debug mode, correct? I am wondering
if there is ever a case where a vnode would not have a free block
entry that we would have to catch in release mode?"