common: Fix COMPILE_ERROR_ON
Allow COMPILE_ERROR_ON outside function scope.

The ones that were present in ASSERTs in out.h
must work with non compile time constants as well,
so the original form is kept there.
The functionlity remained the same, it allows ASSERT(0), and
fails on ASSERT(1):

```
error: size of unnamed array is negative
  ASSERT(1);
    ^~~~~~
```