common: Fix COMPILE_ERROR_ON
Allow COMPILE_ERROR_ON outside function scope.

The ones that were present in ASSERTs in out.h
must work with non compile time constants as well,
so the original form is kept there.
The branch on __builtin_constant_p is removed, as it
served no purpose. The 'if' branch is a runtime feature,
and the COMPILE_ERROR_ON is a supposed to be a compile
time feature, i.e.  the check is performed at compile time
wether the branch will be taken at runtime or not. The
functionlity remained the same, it allows ASSERT(0), and
fails on ASSERT(1):

```
error: size of unnamed array is negative
  ASSERT(1);
    ^~~~~~
```