preload: don't use syscall_no_intercept in error handling
There's not much point in using syscall_no_intercept in pmemfile,
because we don't need to be so highly isolated from libc as is
the case for syscall_intercept.

So instead of building error message with SYS_write just use fprintf.
Also rename exit_group_no_intercept to exit_with_msg.