posix: reserve some space inside of the inode
Only 568 bytes of inode is used to store file data.
This is enough to store . and .. in a directory inode.

Short symlinks are stored directly in an inode.
Longer symlinks (longer than 568 bytes) are stored outside.
Symlink length is now limited only by the largest allocation class.

Empty directory is now reported to have size of 4096 bytes.