antool: print all syscalls to analysis log by default
Only syscalls with a path or a file descriptor among arguments
were printed to the analysis log so far. This patch fixes that
and makes that all syscalls are printed by default now.

However it has to be turned off in some tests that use threads,
because during threads' joining syscalls appear randomly,
so they cannot be checked using match files.