tests: disable rw.failed_write under pmemcheck
This test tries to create file bigger than pool size and expects this
operation to fail.
Some time ago this test was quick under pmemcheck because we tried to
allocate one huge block.
Right now blocks are limitted to 2MB, so we try to allocate this huge
file in 2MB blocks, which means much more work and pmemcheck doesn't
deal very well with long transactions.

Just disable it under pmemcheck.